// components/typing.d.ts
import { GLTFLoader, GLTF } from "three/examples/jsm/loaders/GLTFLoader.js";

export type DroneGLTF = GLTF & {
    nodes: {
        pCube43: THREE.Mesh,
        pCube43_BlackGlass_0: THREE.Mesh,
        pCube43_MetalShiny_0: THREE.Mesh,
        pCube43_Metal_0: THREE.Mesh,
        pCube43_Rubber_0: THREE.Mesh,
        pCube43_lambert1_0: THREE.Mesh,
        pCube43_lambert4_0: THREE.Mesh,
        pCube43_pasted__Eyes_0: THREE.Mesh,
    },
    materials: {
        BlackGlass:THREE.MeshStandardMaterial,
        Metal:THREE.MeshStandardMaterial,
        MetalShiny:THREE.MeshStandardMaterial,
        Rubber:THREE.MeshStandardMaterial,
        lambert1:THREE.MeshStandardMaterial,
        lambert4:THREE.MeshStandardMaterial,
        pasted__Eyes:THREE.MeshStandardMaterial,
    }
}